% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data()
}
\value{
a Gun murder dataset from FBI reports, it also contains the population of each state.
}
\description{
retrieve a dataset with state (US state), abb (abbreviation of US state), region (geographical US region), population (state population 2010), total (number of gun murders in state 2010)
}
\examples{
get_data()
}
