% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1regression.R
\name{l1regression}
\alias{l1regression}
\title{l1regression}
\usage{
l1regression()
}
\value{
the results of the linear regression
}
\description{
Performs a minimum absolute residual (L1) regression
}
\examples{
l1regression()
}
