% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1regression.R
\name{l_regression}
\alias{l_regression}
\title{l_regression}
\usage{
l_regression()
}
\value{
the results of the linear regression
}
\description{
Performs a minimum absolute residual (L1) regression
}
\examples{
l_regression()
}
