% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_rate.R
\name{min_rate}
\alias{min_rate}
\title{min_rate}
\usage{
min_rate()
}
\value{
print the state with the lowest rate, in case this rate is below 0.5, otherwise print "No state has murder rate that low"
}
\description{
check if there is a state with the rate lower than 0.5
}
\examples{
min_rate()
}
