% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{rate}
\alias{rate}
\title{rate}
\usage{
rate(data = NULL)
}
\arguments{
\item{data}{dataframe. If \code{NULL} (the default) selects murders dataset.}
}
\value{
a list of murder rates (in 100000) for each state
}
\description{
computes the rate of crimes for each state in 100000
}
\examples{
rate()
}
