% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_by_region.R
\name{rate_by_region}
\alias{rate_by_region}
\title{rate_by_region}
\usage{
rate_by_region()
}
\value{
graph showing the rate by region
}
\description{
evaluates the rate by region
}
\examples{
rate_by_region()
}
