% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{regression}
\usage{
regression(data = NULL)
}
\arguments{
\item{data}{dataframe. If \code{NULL} (the default) selects murders dataset.}
}
\value{
a list defining the regression
}
\description{
Performs a Minimum Absolute Residual (L1) Regression
}
\examples{
regression()
}
