% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{correlation}
\usage{
correlation()
}
\value{
Graph showing the correlation
}
\description{
evaluate the correlation between the rate and the total population
}
\examples{
correlation()
}
